Fix string literals inside params
Use the same logic we use inside `<Translate>`.

Note: There is an edge case that is very unlikely to happen in any real
application, but if a param's content did have leading/trailing
whitespace, it is now properly stripped during extraction, so the pot
file will change when extracting the strings again.